BF: rerun: Guard against empty result list
There are two cases where we can end up with an empty list of
results: (1) the user specifies --since='' but there are no run
commits in REVISION's history and (2) the user specifies a range
A..B (i.e., --since=A B) that doesn't have any commits.

In most cases, the downstream code will process an empty result list
fine, but, if --onto is specified, we'll fail with an index error.
Abort if we encounter an empty list to prevent the index error as well
as provide more useful feedback to the user.