RF: rerun: Use a more general approach to decide "skip" or "pick"
If we assume a linear chain of commits, we can simply partition the
list into those that are an ancestor of the start point ("skip") and
those that aren't ("pick").

This logic breaks down in the non-linear case, so we need another
approach to support rerunning merges.  Instead, skip the commit if it
is reachable from the current head or the start point, and cherry pick
it otherwise.  Note: This approach assumes the commits are yielded
according to --topo-order.

A downside of this approach is that the output of --report is less
precise: we say "skip-or-pick" because we don't know yet which it will
be.