RF: rerun: Refuse to run root commits
We abort if --since='' and --onto='' are given and there is a run
commit at the root of the history.  The motivation is that we're very
unlikely to encounter a repo that has a run as the first commit, so it
isn't worth the complication to support, particularly when a branch
isn't specified.

When rerun is extended to handle merges, it will be possible to have
multiple roots if there is a merge of unrelated branches.  To make the
logic and processing easier, refuse to run all root commits.