RF: rerun: List revisions with --topo-order
In preparation for supporting rerunning merges, use the --topo-order
flag so that commits on the same line are grouped.  This will make it
easier to process the revisions because we know that if we encounter a
commit on a different line, we can switch to processing that line
without keeping track of the previous one.