ENH: rerun: Support rerunning history with merges
Many of the pieces are in place with the preceding commits.  The main
thing here is to create a map of old bases to new bases so that we can
re-execute a commit on the right base.  We make branch management easy
by not doing it.  Instead, we just work in a detached state and, if
the user specifies a branch, we point that branch to HEAD at the end.

We don't yet handle merge conflicts.  Those, along with cherry pick
conflicts, will be addressed later.

Closes #2093.