TST: Mark test_rerun_cherry_pick as a known direct mode failure
Before 90d13c965, this test passed because files were added to git
rather than annex.  This is no longer the case, which means that git
refuses to let us check out another revision for the rerun because the
annex files are seen as type changes.  Thus this is a legitimate
direct mode failure that was masked by an issue with add().