TST: run: Make some status checks more specific
In direct mode, dataset.diff sees all annex files as type changes.
Because of this, some of the blanket "all status X" checks in
test_basics are failing in direct mode.  Where we expect all
"notneeded"s, there's an "ok" because the typechange diff result.
Where we expect all "ok"s, there's a "notneeded" because of an add
call on a typechanged annex file that is already in the dataset.

These weren't failing before 90d13c965 because the files were added to
git instead of the annex.

Adjust the checks on the result list to be more specific.