TST: Mark test_remove_more_than_one as a known direct mode failure
Before 90d13c965, this test passed because files were added to git
rather than annex.  This is no longer the case, and git fails with

  Failed to run ['git', '-c', 'receive.autogc=0', '-c', 'gc.auto=0',
  '-c', 'core.bare=False', '--work-tree=.', 'rm', '-r', '--', 'two',
  'one'] under [...]. Exit code=1. out= err=error: the following files
  have local modifications:
      one
      two
  (use --cached to keep the file, or -f to force removal)

Perhaps we should run 'git rm' with --force when dataset.remove is
called with check=False, but, at any rate, this is a legitimate
failure that was masked by an issue with add().