RF+BF: gitrepo: Simplify return value of get_changed_files
get_changed_files returns a dict of files mapping to status codes, but
it has two bugs.  It would break in the inadvisable and rare situation
where someone includes a newline in a file name.  It's not a priority
to handle such file names, but we might as well if we can.  More
importantly, the diff call doesn't disable rename detection, so it
will incorrectly parse a R or C entry, which includes two file names.

One solution would be to properly parse the full output of "-z --raw",
dropping "--name-status".  (Or better still, use git-diff-files and
git-diff-index, which aren't influenced by the diff.renames variable
like the porcelain git-diff is).  But no callers currently use status
codes, so let's simply use "--name-only -z".  Also, interface.diff
already does parsing of the raw format, so it makes sense to keep this
low-level interface as simple as we need it to be, and try to use
interface.diff wherever we can.