TST: Skip test_install_recursive_repeat on Windows
There are a couple of things going on here.  Before 2e39ce546, this
was skipped on AppVeyor as a known direct mode failure .  If it
wasn't, it would have failed with "file used by another process
error".  It's happening in a codepath of rmtemp not covered by
a670ddd55 (Add explicit call of gc and loop inside rmtemp(),
2015-05-06).

However, even if we avoid the above error by not removing the test
directories, the ok_clean_git call fails.  It seems to stem from the
git annex status calls in annexrepo.get_status.  These repos look
clean on manual inspection.  'git annex status' reports some of the
some of the Git tracked files (like .datalad/config) as modified, but
the file on the system and the one from the Git tree are identical.

Punt on this for now since it isn't directly related to this series.