BF: downloaders: Fix needs_authentication assignment
The downstream code makes the distinction between False and None, but,
with the default value of None for self.credential, there is no way
for the code to evaluate to False.

Assume that the intention was for needs_authentication to be set to
requires_authentication (True/False) when there's an authenticator.