CLN+RF: Remove a seemingly inaccessible codepath
I can't find a way to trigger this with the tests.  When a downloader
class is initialized with an authenticator but no credentials, the
downloader refuses to continue unless the user enters credentials.  So
it seems impossible to execute a codepath that assumes that there is
an authenticator but no credentials.