ENH: GlobbedPaths: Expand sub-patterns if full pattern fails
Chop off the tail of the pattern and stop at the first sub-pattern
that matches.  By doing this, we make it possible for the expand()
caller to match an uninstalled subdataset that contains a path,
install that subdataset, and repeat.

Note that if we weren't interested in supporting globs in the leading
path, we could avoid _get_sub_patterns and simply try to install the
leading path (if any) and then reglob.