BF: run: Support globbing in uninstalled datasets
In order to resolve globbed paths within a subdataset, the subdataset
needs to be installed.  To do so, use the following procedure: install
the subdirectories returned by the longest *matched* pattern and then
glob again, stopping when no new paths are globbed.

This relies on two features of the install call: (1) it will quietly
ignore subdirectories that aren't datasets and (2) it will install the
dataset when given a path to a directory within the dataset.

Punt on handling rerun_outputs.  At the moment the datalad.diff call
in rerun.py will fail if a subdataset isn't installed, and we need to
address that first.  (There's also the outstanding issue of the
subdataset state not being reset on rerun.)

The new tests pile on to the Windows incompatibility, but at this
point test_run.py needs a large-scale rewrite.

Fixes #2789.