BF: run: Escape each input/output before formatting command string
Our processing of inputs and outputs is broken for paths with
spaces.  Take the inputs "foo blah.txt" and "bar.txt", and the run

  $ datalad run --input="a" --input="b c" CMD

If CMD is a string (e.g., "./script {inputs}"), then the formatted run
command is ultimately "./script a b c" rather than "./script a 'b c'"
because the individual inputs aren't quoted.

On the other hand, if CMD is a list (no quotes), the list is first
converted to the string form of "./script '{inputs}'" (see 9c3a65e3d)
and ultimately "./script 'a b c'".

Quote each input and output so that paths with spaces work with the
string form.  This won't work with the list form because {inputs} is
surrounded by quotes, but at least one form will work reliably.