BF: annexrepo.file_has_content: Fix check for unlocked files in V6
To check whether a file has content, we either look for a link with
the expected form (allow_quick=True) or check if "git annex find"
returns the name.  The first approach fails for obvious reasons on
unlocked files in V6.  The second approach fails because "git annex
find" on an unlocked V6 file---even a modified one---returns the name,
unlike the V5 call which fails.

Handle unlocked files in V6 with the following rule: report that an
unlocked file has content only unless *unstaged modifications*.  This
is probably the closest we can get to consistent behavior between V5
and V6.  It's possible that modified content as already been added by
the clean filter into the annex, but this doesn't seem worth dealing
with.

Note: If we're in V6, we send everything through the slower annex
check, but we could still check locked files with the quick check,
which would probably improve performance.  Punt on this for now
though.

Fixes #2845.