BF: annexrepo.find: Ignore batch if a directory is present
If 'git annex find --batch' receives a directory, it always returns
empty output, in contrast to plain 'git annex find', which returns the
available annex files in the directory.

Currently the callers in test_aggregate.py are the only ones that pass
a directory, and they use batch=False, so they aren't affected by
this.