BF: annexrepo.find: Return dict to accommodate directories
'git annex find' will return the list of available annex files if
given a subdirectory as an argument.  We currently leave this as a
single, unparsed item.

We can't simply expand it because find() returns a list with a single
item for each input.  Instead let's change the return value to be a
dictionary.  A nice side-effect of this is that it makes find's return
value more closely match info's.