BF: annexrepo.is_under_annex: Disambiguate files from remotes
If a file name matches a remote name (or "here"), 'git annex info'
will interpret "here" as well it as a remote, not a file.  Prepend
"./" to prevent this from happening.

This should probably be taken care of at the leave of info() itself,
but then we have to worry about whether we should return a filename
that is different than the input one.  Also, annex.info has pretty
strict checks, so a caller will know that it is happening.  We don't
add a test for this now, but an upcoming commit will fix V6 checks for
unlocked modules, and then test_partial_aggregation exposes this issue
because it has a file named "here".