BF: do not detect RIs which start with / as ssh urls
The reason is that on Windows, e.g. under Anaconda shell, path would
look like /C:/DIR/SUB..., and split by : into two components which we
considered sufficient to assume that it is an SSH url.  Clearly not the
case here, so we do not detect any RI which starts with / in the first
part as ssh RI