BF: Reevaluation of config and repo properties
Reevaluate properties when physically underlying repo changes, ceases to
exist or comes into existence.
Note, that Dataset is supposed to represent a path reference to a
repository that may not even exist yet, whereas *Repo classes are
supposed to represent the actual repository. This leads to the Repos
being flyweights wrt to realpath and Dataset being flyweight wrt the
unresolved path!