BF(TST): realpath path before figuring out which repo it belongs to
Surprisingly ran into it only within https://github.com/datalad/datalad/pull/2898
despite all our testing with symlinked TMPDIRs on travis.  This time I have added
a dedicated unittest to trigger that failure.

The rationale for using realpath there is because GitRepo already realpath"ing
the path to be able to return the singleton Repo for the same path