CLN+TST: Drop known_failure_v6 decorator
git-annex now has a repository version 7.  We should update
known_failure_v6 for this because, if something fails in v6, it very
likely fails in v7.  But this is no longer used for any tests (aside
from test_AnnexRepo_get_submodules, which doesn't actually do
anything).  Let's just remove it.  If we ever need it, we can easily
revert this and adjust known_failure_v6 for v7.