RF: Allow for customization of change recording in `run`
This is admittedly a bit clunky. An alternative way of doing this
would be to RF the `add()` call into the main command class, and
have a helper method deal with it. In this scenario, it could simply
be overwritten in a derived class, instead of having to introduce
another argument. What would you prefer @kyleam?

Why is this needed? ATM `run` performs a hard-coded `add(recursive)`,
and this adds the flexibility to tailor this behavior (non-recursive add,
or more than just adding).