TST+BF: download_url: Don't try to infer dataset if save=False
When a dataset isn't explicitly specified, download_url will try to
infer one from the current directory.  However, if save is False,
there's no reason to do so.  Trying to infer the dataset in these
situations generally won't hurt, but doing so may unnecessarily fail
in the save=False case if the current directory is something that may
cause issues when used as a dataset (e.g., a git worktree).

Note: This was prompted by a download_url test failing when ran from a
worktree (gh-3028).  The most direct solution is to use chpwd() to get
to a non-dataset directory in the cases that we call
api.download_url(dataset=None), but that wouldn't help for callers
outside the test.  Another would be to catch the repository exception
around the require_dataset() call, but I've found that, when calling
`datalad download-url` from a worktree, the exception may be raised
downstream rather than at the require_dataset() call.

Fixes #3028.