RF: run: Rework normalize_hits helper
Convert normalize_hits to operate on a single hit rather than a list
of hits.  Working on a list makes sense to avoid repeating the
sorted() call, but the normalization logic is going to get an extra
condition, and it will be easier to work with a single hit.  We could
also add another helper function to avoid repeating sorted(map(...)),
but that's an overkill.