BF: run: Preserve current directory prefix for inputs/outputs
When globbing inputs/outputs, we normalize them by making them
relative to the directory the command will run in.  In the eyes of
os.path.relpath(), the "./" on "./path" is superfluous, and it drops
it.  This is usually OK, but a datalad run caller may rely on it to
format a script call:

    datalad run --input=./script {inputs} ...

Fix this by returning the hit unchanged if the normalized version
looks the same, sans the current directory prefix.  It's possible that
there is an edge case that this misses, but it's simple, should catch
the main cases, and shouldn't result in any false positives.

Closes #3034.