BF(py2): run: Ensure that the command is unicode
The tests added in gh-3035 tested the Python API case, where the run
caller passes in unicode strings.  But when called from the command
line, the command will be a byte-string in Python 2.  Ensure that it's
unicode.  Do this in normalize_command() because otherwise we need to
repeat the list/string test elsewhere and outside callers (like
containers-run) would need to apply a similar py2-compatibility fix.