RF: run: Do not force byte-encoding of commit message
This is like the patch from gh-3031 except that it still encodes the
message before writing it for py2 compatibility.  This approach should
still address the failure on windows (gh-3030).

Remove the assure_bytes() call not only because it is causing issues
on windows but also because conceptually there isn't a reason to
encode the message here.  It will ultimately be passed to GitRepo or
AnnexRepo's commit(), which shouldn't and don't require a byte-string
message.  Current tests indicate there is no issue, and, if one
arises, it should be fixed in {GitRepo,AnnexRepo}.commit().