BF: run: Use 'datalad add' in failure instructions
If a run command has a non-zero exit, we don't save the outputs and
instead instruct the user how to save the outputs given the failure is
expected.  The instructions say to use 'datalad save -r .', but there
are cases where this is not sufficient to add an output file (e.g., if
a file is added to subdataset).  Update the instructions to use 'add'
instead of 'save', which is what run actually uses (by default)
underneath.

Also add '-d .' to the instructions to match run's dataset.add(),
reducing the possibility of discrepancies like the one that cropped up
in gh-2276.  There are still cases where run's dataset.add() is not
sufficient (a known one is gh-2454, which is addressed in
datalad-revolution), but at least the command in the instructions will
match run's behavior.

Closes #3072.