RF: save: Ensure that message file content is decoded
Under Python 2, we pass a bytestring to commit(), but the intention
was to pass a unicode string, like we do under Python 3.  Passing a
bytestring as the message doesn't seem to cause any issues, at least
according to test_save_message_file, but we should still do so for
consistency across Python versions.