RF: save: Use ValueError for --message/--message-file conflict
When save is called from the command line with both --message and
--message-file but without -d<dataset>, we get two error messages, the
intentional error message as well as one about a key error:

  [ERROR  ] Both a message and message file were specified
  [ERROR  ] 'path' [utils.py:default_result_renderer:503] (KeyError)

This is because we pass refds_path as the path to get_status_dict(),
but refds_path is None in the above scenario.

Instead of worrying about properly populating the status dict info,
which isn't particularly relevant for the --message/--message-file
conflict, just raise a plain exception.  Doing so is in line with
what's done in other interfaces (e.g., aggregate raises a ValueError
when it receives an unknown update mode).