ENH: run: Strip leading "--" to allow command disambiguation
The command is collected with argparse's REMAINDER.  In the context of
DataLad's command-line parsing, this usually works fine but (1) has a
known issue with parsing at least one special flag, --version, and (2)
doesn't allow the command to start with an option, which is a common
situation when using runscripts with container-run.  "--" can be used
in both these cases to explicitly tell argparse that the option is a
part of the command; we just need to remove it from the command list.

Fixes datalad/datalad-container#46.
Re: #3067