Centrlize handling running commands with long files list in _run_command_files_split
Before now we were doing that only in _run_annex_command_json but
the problem is present also for other commands (e.g. add) which
do not use --json mode of annex, or are pure git commands (e.g. git add).
_run_command_files_split(func, files, ...) helper would invoke func
multiple times on batches of files, and code is refactored to use it
for annex and git commands invocations.  I might have missed additional
places where it should be used.

Also I am not sure if it would not interfer with custom index file, and
most likely would cause multiple commits instead of a single one for
commit calls. But imho it would already be better than just crashing