TST: protocol: Don't use GitRepo() to test callable
protocol.py distinguishes between external commands and callables.  To
test the "callable" case, we create a GitRepo, assuming that this will
lead to gitpy.Repo.init() being executed through the runner.  But in
-revolution, RevolutionGitRepo overrides _create_empty_repo() and uses
the external command 'git init' as of 8b9d65a (ENH: Expose git-init
options, create without GitPython, 2018-10-29).  As a result, these
protocol tests fail if we substitute RevolutionGitRepo for GitRepo.

Use os.mkdir to prepare for the absorption of RevolutionGitRepo into
GitRepo.