ENH: ssh: Support configuring an identity file
This enables callers to use sshrun with non-standard identity files.
For example, in the ReproMan project, we want to use DataLad commands
with an AWS EC2 instance.  In this case, the host will not be in
.ssh/config, and the key files are usually not in the default
location.

A plain environment variable would do for the above usecase, but add
it as a variable in common_cfg.py for the visibility and
documentation.