TST: gitrepo: Prepare test for revolution's _create_empty_repo()
RevolutionGitRepo overrides _create_empty_repo() and uses the external
command 'git init' as of 8b9d65a (ENH: Expose git-init options, create
without GitPython, 2018-10-29).  This changes some details of the
recorded protocols because _create_empty_repo() no longer sets the
_repo attribute in __init__().  This means that the commit() call
leads to _repo being set, and this adds another item to the protocol.

We're about to absorb RevolutionGitRepo, so adjust our tests to not
make such strict assertions about these internal implementation
details.  The point of this test is simply to check that GitRepo is
using the supplied custom runner.