Prepare for first preview release of 0.12 series
I see no point of adding tons of changelog entries for individual
pre-releases (not even a real RC yet). Instead, I started 0.12.0 entry
so we can fill it up with what is to come.

I would have prefered to call this 0.12.0pre1, but setuptools insists
on "normalizing" the version to 0.12.0rc1 -- so we better use that right
away.