BF: Fix LORIS token authenticator type with Python3
The HTTPResponse object returned by urlopen under Python3 is not
compatible with json.load(!) because read() returns bytes, while json.load
is a thin wrapper around json.loads (which expects a str). This stores
the response as a str in a temporary variable and converts it to a string,
before calling json.loads to prevent a type error when used under Python3
while trying to authenticate against a LORIS instance.