DOC: PR template: Convert to an instructions checklist
Make it a set of instructions to be deleted rather than mixing text
that is intended to be kept (like "This pull request fixes") with
instructions that should be deleted (like the Zenodo stuff).  The
advantage of this is that submitters don't have to try to figure out
what they are expected to keep.