Do not decode fetched content when retrieving status
When annexificator is used in a datalad-crawler pipeline, it attempts
to call get_status on the URL to see if the file has changed. get_status
attempts to use _fetch, which attempts to decode the content under Python3.

This is invalid for content that isn't utf-8, and get_status was only
trying to verify that it wasn't receiving a login or failure page or similar.
_fetch, confusingly, converts the UnicodeFormatError into a generic DownloadError
(despite the fact that the download succeeded, but conversion to a string failed.)

This teaches _fetch a "decode" parameter that toggles whether or not it should
even attempt to decode. In the case of get_status, the content is discarded,
so the decode serves no purpose.

NB. fetch (without the _)'s documentation claims that it doesn't decode, but
it calls _fetch (which currently always decodes.) This can new parameter can
eventually be used to fix that, but currently trying to fix "fetch" results in
a json.loads error from a mysterious part in the code if fetch is updated to
not decode as documented.