TST: travis: Widen selector for extension runs
All of the extension runs use 'NOSE_SELECTION_OP=' and
'NOSE_SELECTION="integration or usecase or slow"', restricting the
runs to tests _with_ these attributes.  This presumably wasn't the
intention because it results in very limited runs.  For example, the
crawler run executes only three tests [0].

Adjust NOSE_SELECTION_OP and NOSE_SELECTION for each extension run so
that all tests are executed.

[0]: https://travis-ci.org/datalad/datalad/jobs/503126961