BF: workaround for the check in direct mode
Relies on the fact that ls-files lists files in the top repo,
which unlikely to contain exactly the same files in the directory we
are trying to init in.  Then in direct mode we would result in simply
skipping this safe-guard.  Given unknown amount of direct mode usage
and its imminitent death of support in datalad - it is a reasonable
price to pay.

For indirect mode repos this check would rarely get executed since
unless inthe cases where to blow, there would be no files reported.