BF: Fix PWD to match provided cwd in Runner
We do take care about tuning up PWD while chdir etc.  But when we run
external process and set its cwd, we left PWD unchanged if was provided
in (likely) inherited and tuned environment.  This fix would change
PWD to match provided cwd, if it is known to the environment and cwd
is set.

I was not bold enough to always get an env (if not provided) and change it
there, but may be that is what we will need to end up doing at some
point.

This should resolve the failing test in
https://github.com/datalad/datalad/pull/3184
caused by underlying datalad process, which is started from a temp directory
while we provide it with a now non-matching PWD:
https://travis-ci.org/datalad/datalad/jobs/501525590#L4325