RF(TST): travis - do not call  pip install -e .  with sudo
Originally done in e64286864f38883b119532fe37e906efe312c7a9 but rationale
is not 100% clear for "system-wide" purpose here.  As a side-effect it
causes pip to complain:
  The directory /home/travis/.cache/pip/http or its parent directory is not owned by the current user and the cache has been disabled....
which (I believe) in turn causes travis to not fold that output for
more convenient overview of the log.

Hopefully a regular  pip install -e .   would be "wide" enough to verify
correct installation, and would preclude such warnings