ENH: run: Add 'tmpdir' placeholder
For some commands, it is useful to have a temporary directory.  For
example, one can be used as the source component of a Singularity or
Docker mount point.  Expose a temporary directory with the '{tmpdir}'
placeholder so that users do not need to embed $(mktemp -d) within the
command or write a command wrapper.

Re: datalad/datalad-container#50