RF: Create directory structure needed for absorbing rev commands
Following the directory structure proposed by gh-3192, all of the
about-to-be-absorbed commands would got to core/local/.  The
disadvantage of using this structure right now is that, because other
datalad components haven't been migrated, it could be confusing for
someone inspecting the repository.  The advantages are that the rev
commands go into a separate space for now, and we won't need to move
them again assuming we migrate to the proposed layout.