MNT: run: Drop deprecated --rerun flag
This has been deprecated for a while; it's time to retire it.  Doing
it _now_ is useful because the next commit will make rev-run a pure
alias for run.  If --rerun was still around, we'd either have to
duplicate it in rev-run's __call__ or have a slight difference between
run and rev-run.