BF(workaround,ugly): allow to pickup metadata from superds if ds lacks metadata for a file
Well, largely it is just the issue of semantic difference whenever -d is provided
or not.  Without -d context is moved into the ds where the path "belongs", and if
it is the subds path, context shifts inside that subds, and no refds/parentds is
set.  That makes it impossible cleanly (thus I started adding deduction of the
parentds right in the metadata code which is wrong) to handle situation agnostic
of either ds was provided or not