Provide a warning when a regex in a provider isn't valid
When a provider has an invalid regular expression, datalad
currently provides obtuse error messages like
"bad character range [re.py_compile:251] (error)".

This catches any exceptions while compiling the re to explicitly
warn the user that there is an invalid regular expression in the
provider.

Unfortunately, there doesn't seem to be any way to provide the
path to the provider config file at this point in the code, so we
simply give the provider name.