ENH: ui: Drop progress bars if not attached to tty
If is_interactive() returns false, we select the ConsoleLog backend.
This backend displays progress bars, but, in most non-interactive
cases, the progress bars are noise, so let's instead set the backend
to SilentConsoleLog.

If there turns out to be a need for progress indication in these
cases, we can expose a configuration variable for the backend (right
now we have only datalad.tests.ui.backend) and/or use log_progress()
more widely so that we show log messages instead of a progress bar if
not attached to a tty.

Closes #3092.