ENH: Record a dataset's ID in a superdataset on `add/save` (fixes gh-3303)
Otherwise we have incomplete identity information on all dataset
components (i.e. we only have the state of a subdataset, but not its
persistent ID). This is an issue during metadata extraction, where we
would require subdatasets to be around in order to be able to give a
full report.

We already ensure that create(force=True) does not alter a dataset's
ID, so getting out of sync would require manual fiddling with DataLad's
internal config. Sure be safe...