ENH: utils: Make py2 Pathlib objects work with unicode()
It's tempting to say text_type(pathobj) and hope for it to do the
right thing, but that will fail.

    python -c "from pathlib2 import Path; print(unicode(Path(u'β')))"
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
    UnicodeDecodeError: 'ascii' codec can't decode [...]

Even if a unicode string is passed in, pathlib2 stores it as bytes and
doesn't define a __unicode__ method.  Define a __unicode__ method for
our pathlib2 objects so that we can call text_type() on them and get
back unicode.