BF: create-sibling - do NOT install post-update hook unless web UI
AFAIK post-update hook is useful only for updating web-ui information and running
"git update-server-info" for dummy web git backend.
It also calls "git submodule update --init" which might be generally not desirable.
Running that hook requires datalad on remote end, and could be quite time consuming
for any sizeable repo.  And there is no reason for that unless there is web-ui
somewhere in the super-dataset.

What this one breaks is creating missing remote siblings on a deployment WITH web ui,
while asking to inherit settings.  Next commit should work on that to bring it back.

Also we call out to all the hooks upon create-sibling at the end, but if there is no hook,
it should just skip that.