RF: do not run submodule update --init  in the post-update hook
All tests seems to pass, so I guess whatever issue was attempted to be solved
by this was resolved  by introducing correct traversal of the datasets
(create on going down, publish on going up), so no submodule update --init
call should be needed.

Side-effect it was causing is instantiating repositories on the remote
which might actually just be not a part of the dataset locally but just pointing
to remote urls (e.g. to github).