BF: do not blow while reconfiguring from superdataset
It is often desired to assure that we have consistent hierarchy
reflecting current one on the remote site.  But currently --inherit
would blow for the topmost super dataset since there is nothing to inherit
from.  With this change - we just issue a warning, and do not blow.
_DelayedSuper (not sure how useful it is really) will just report .super=None
in such cases so it could be verified that we can or cannot inherit
from it