BF: status: Provide special treatment of "." path
When status receives "." as the path, rev_resolve_path() resolves the
path to the dataset (either to the one explicitly provided as an
argument or to the one inferred from the current directory).  So,
using "." means we get the reference dataset as the path.

If a call includes an explicitly specified dataset _and_ "." as a path
argument, whether the call succeeds or fails depends on whether the
specified dataset is a subdataset.  When the reference dataset isn't a
subdataset, status queries within the dataset.  When it is a
subdataset, the call fails with a message saying the subdataset is not
underneath itself:

  $ datalad status -d. .
  [ERROR ] dataset containing given paths is not underneath the
  reference dataset [...]

Even though the desired behavior could be achieved by dropping "." as
a path argument, let's make using "." as the path argument work
because

  * it is consistent with the non-subdataset case, and it won't be
    obvious to a caller why it works in one case but not the other

  * based on the use of "." as a path argument in other places in
    datalad, it's be reasonable for a caller to expect, say,
    subds.status(".") to work

As shown by the added run test, this fixes a regression where calling
'datalad run' failed from within a subdataset, which was introduced by
eba3ccfd6 (BF: run: Save results with rev-save, 2019-03-18) because
the call to rev_save() retained the path="." that is necessary for the
add() call.