BF(py2): pathlib: Consistently use text_type for conversion
Our code makes heavy use of str(pathobj), but under Python 2 that can
be problematic anytime the path has non-ascii characters.  Convert all
the str(pathobj) instances to text_type(pathobj).  Note that this
depends on the Python 2 pathlib objects working with unicode(), which
is the case as of f0822dcdf (ENH: utils: Make py2 Pathlib objects work
with unicode(), 2019-04-08).

Most of this is straight substitution or u-prefixing, but there is an
assure_unicode() call in the result renderer of status that's a bit
surprising.  The path would be expected to be a unicode string here.
On PY2 it is not, because interface.utils encodes the result values on
PY2 before they get to custom_result_renderer() as of
616e4e6ed (Unicode fixes for test_add (PY2), 2018-04-13).

Note that we're using OBSCURE_FILENAME in the tests.  That sacrifices
compatibility with older Windows (see gh-2929), but, without it or
something that probes the system, the current appveyor run will fail:
https://ci.appveyor.com/project/mih/datalad/builds/23828771/job/w6wf4myq5o6nyna6#L696