ENH: ansi_colors: Add force parameter to color_word()
Make it possible to force the use of colors so that color_word() can
be used in log.ColorFormatter, which should obey the setting of its
use_color attribute.

Using color_word() in ColorFormatter arguably simplifies the code and
improves readability, and it will avoid repeating the conditional when
we start mapping a false color value to "no color".