ENH: ansi_colors: Support "default" color
The INFO level name is set to white in LOG_LEVEL_COLORS.  For
terminals with a dark background, this works as the default,
unremarkable color.  For terminals with a white background, the user
sees "[       ] ...".

Teach color_word() that a false value means "don't color", and change
INFO's color to None so that the default terminal color is used.