ENH: status: Use None, not white, as default color
The result renderer colors the state white if the state has not been
assigned a color in STATE_COLOR_MAP.  This is unfortunate for users
with a white terminal background.  A deleted file, for example, would
appear as

  $ datalad status
           : foo (file)

To instead use the default terminal foreground color, set the color to
None if the state isn't in the color map.  color_word() recently
learned to not paint the word if it receives a false value as the
color.