BF(py2): Avoid instantiating GitPython repo with unicode path
Under Python 2, passing a unicode path to GitPython's Repo fails as of
GitPython v2.1.11, specifically 7f08b77 (Allow pathlib.Path in
Repo.__init__, 2018-07-12).

In an attempt to handle pathlib objects, GitPython converts the path
with str() if it isn't already a str.  Under Python 2, if a unicode
object is passed in, str() is called on it, leading to a possible
encoding error.  For Python 2, work around this by passing in an
encoded path.